requirements.txt
# Core dependencies
python-dotenv>=0.19.0
requests>=2.26.0
numpy>=1.26.4
pandas>=2.2.3
datetime>=4.3
anthropic>=0.5.0
tenacity>=8.0.0
pyarrow>=21.0.0

# Browser automation
selenium>=4.1.0
webdriver-manager>=3.5.2

# Data processing and analysis
matplotlib>=3.5.1
scikit-learn>=1.0.2
statsmodels>=0.13.2
scipy>=1.8.0

# Technical indicators and analysis
ta>=0.10.1

# Machine learning - TensorFlow for Apple Silicon
tensorflow-macos==2.16.2
tensorflow-io-gcs-filesystem==0.37.1

# Database
SQLAlchemy>=1.4.32

# Utilities
tqdm>=4.62.3
typing-extensions>=4.0.0

# Google API (from sheets_handler)
google-auth>=2.6.0
google-api-python-client>=2.40.0

# ============================================================================
# M4 OPTIMIZATION PACKAGES - NEW ADDITIONS
# ============================================================================

# Polars - Ultra-fast data processing (Rust-based, 10-100x faster than pandas)
polars>=1.16.0

# Numba - JIT compilation for mathematical functions (3-10x faster)
numba>=0.60.0

# aiohttp - Async HTTP requests for faster API calls
aiohttp>=3.11.0

# ============================================================================
# PERFORMANCE MONITORING (OPTIONAL)
# ============================================================================

# Memory profiling and system monitoring
psutil>=6.1.1
memory-profiler>=0.61.0